// src/app/shared/models/typeBandeFrequenceList.ts
/**
 *
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TypeBandeFrequenceList {
    readonly id: number;
    libelle: string;
    categorie_produit : number;
}